Demo cleanup

+++ Foundation/Lib/source/hash/CRC32Util.cpp
@@ -12,5 +12,5 @@ std::uint32_t crc32Calculate(const std::uint8_t *data, std::uint32_t size)
 std::uint32_t crc32Calculate(const std::string &data)
 {
   return crc32Calculate(reinterpret_cast<const std::uint8_t *>(data.data()),
-                        data.size());
+                        static_cast<int>(data.size()));
 }
