+++ Graphics/Lib/source/collision/CCollisionSystem.cpp
@@ -12,7 +12,7 @@ unsigned int CCollisionSystem::getNewGroupId()
   // Add new collision group
   m_entities.push_back(std::list<CCollidable *>());
   // Return the group id
-  return m_entities.size() - 1;
+  return static_cast<unsigned int>(m_entities.size()) - 1;
 }
 
 // Cleanup
